refactor schema publisher to aid testability
