+++ package.json
@@ -41,7 +41,7 @@
     "cypress": "^3.8.2",
     "fetch-mock": "^9.9.0",
     "isomorphic-fetch": "^2.2.1",
-    "jest": "^24.8.0",
+    "jest": "^26.6.3",
     "jest-junit": "^7.0.0",
     "lolex": "^4.2.0",
     "mini-css-extract-plugin": "^0.8.0",
@@ -75,7 +75,12 @@
   ],
   "nodemonConfig": {
     "ext": "js,graphql,yaml,jsx",
-    "ignore": ["*.cyp.js", "**/__tests__/*.js", "test-helpers", "cypress"]
+    "ignore": [
+      "*.cyp.js",
+      "**/__tests__/*.js",
+      "test-helpers",
+      "cypress"
+    ]
   },
   "engines": {
     "node": "14.x"
