+++ packages/tc-schema-publisher/scripts/deploy.js
@@ -1,19 +1,17 @@
 #!/usr/bin/env node
 /* eslint-disable no-console */
-
-const program = require('commander');
 const { join } = require('path');
-const { statSync, readFileSync } = require('fs');
-const schema = require('@financial-times/tc-schema-sdk');
+const { statSync } = require('fs');
+const { SDK } = require('@financial-times/tc-schema-sdk');
 const { sendSchemaToS3 } = require('..');
 
-const pkg = JSON.parse(
-	readFileSync(join(__dirname, '../package.json'), 'utf8'),
-);
+const deploy = async command => {
+	const {
+		schemaDirectory,
+		bucketName,
+		env,
+	} = command;
 
-const action = async (...args) => {
-	const [command] = args;
-	const { schemaDirectory, bucketName, env } = command;
 	try {
 		if (!bucketName) {
 			throw new Error(
@@ -34,42 +32,14 @@ const action = async (...args) => {
 		if (!stat.isDirectory()) {
 			throw new Error('schema directory is not a directory');
 		}
-
-		schema.init({ schemaDirectory });
-		await sendSchemaToS3(env, bucketName);
+		const schema = new SDK({ schemaDirectory, includeTestDefinitions });
+		await sendSchemaToS3(env, bucketName, schema);
 		console.log('successfully deployed');
 	} catch (err) {
 		console.error('Failed to deploy');
 		console.error(err);
-		process.exit(2); // eslint-disable-line unicorn/no-process-exit
+		throw err;
 	}
 };
 
-const help = () => {
-	const template = `
-Example:
-
-  tc-schema-publisher -D ./example-schema -B schema-bucket -v latest
-`;
-	console.log(template);
-};
-
-program
-	.name('tc-schema-publisher')
-	.description('Publish schemas to S3 bucket')
-	.option(
-		'-D, --schema-directory <directory>',
-		'directory to the schema. (default: "process.env.TREECREEPER_SCHEMA_DIRECTORY")',
-		process.env.TREECREEPER_SCHEMA_DIRECTORY,
-	)
-	.option(
-		'-B, --bucket-name <bucket>',
-		'S3 bucket name which you want to upload. (default: "process.env.TREECREEPER_SCHEMA_BUCKET")',
-		process.env.TREECREEPER_SCHEMA_BUCKET,
-	)
-	.option('-E, --env <env>', 'specify publish environment', 'latest')
-	.version(pkg.version)
-	.action(action)
-	.on('--help', help);
-
-program.parse(process.argv);
+module.exports = { deploy };
